==> 9_Hybrid_Recommendation.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

import pandas as pd

train = pd.read_csv('/content/drive/My Drive/BT4222_Group_3_Submission/cleaned_datasets/train.csv')
eval = pd.read_csv('/content/drive/My Drive/BT4222_Group_3_Submission/cleaned_datasets/eval.csv')
test = pd.read_csv('/content/drive/My Drive/BT4222_Group_3_Submission/cleaned_datasets/test.csv')

jobs_score_mat = pd.read_csv('/content/drive/My Drive/BT4222_Group_3_Submission/score_matrices/scoring_matrix_jobs.csv')
implicit_score_mat = pd.read_csv('/content/drive/My Drive/BT4222_Group_3_Submission/score_matrices/scoring_matrix_implicit.csv')
item_cf_score_mat = pd.read_csv('/content/drive/My Drive/BT4222_Group_3_Submission/score_matrices/scoring_matrix_item_based_cf.csv')
ncf_score_mat = pd.read_csv('/content/drive/My Drive/BT4222_Group_3_Submission/score_matrices/scoring_matrix_ncf.csv')
lightgcn_score_mat = pd.read_csv('/content/drive/My Drive/BT4222_Group_3_Submission/score_matrices/scoring_matrix_lightgcn.csv')
cb_score_mat = pd.read_csv('/content/drive/My Drive/BT4222_Group_3_Submission/score_matrices/scoring_matrix_content_based.csv')
user_cf_score_mat = pd.read_csv('/content/drive/My Drive/BT4222_Group_3_Submission/score_matrices/scoring_matrix_ubcf.csv')

### Get 1000 selected reviewer names

reviewer_names = jobs_score_mat.iloc[:, 0].tolist()
print(reviewer_names)

### Remove other reviewers that we are not making recommendations for

final_implicit_score_mat = implicit_score_mat[implicit_score_mat.iloc[:, 0].isin(reviewer_names)].reset_index(drop=True)
final_item_cf_score_mat = item_cf_score_mat[item_cf_score_mat.iloc[:, 0].isin(reviewer_names)].reset_index(drop=True)
final_lightgcn_score_mat = lightgcn_score_mat[lightgcn_score_mat.iloc[:, 0].isin(reviewer_names)].reset_index(drop=True)
final_ncf_score_mat = ncf_score_mat[ncf_score_mat.iloc[:, 0].isin(reviewer_names)].reset_index(drop=True)
final_jobs_score_mat = jobs_score_mat[jobs_score_mat.iloc[:, 0].isin(reviewer_names)].reset_index(drop=True)
final_cb_score_mat = cb_score_mat[cb_score_mat.iloc[:, 0].isin(reviewer_names)].reset_index(drop=True)
final_user_cf_score_mat = user_cf_score_mat[user_cf_score_mat.iloc[:, 0].isin(reviewer_names)].reset_index(drop=True)

print(final_implicit_score_mat.shape)
print(final_item_cf_score_mat.shape)
print(final_lightgcn_score_mat.shape)
print(final_ncf_score_mat.shape)
print(final_jobs_score_mat.shape)
print(final_cb_score_mat.shape)
print(final_user_cf_score_mat.shape)

### Convert Reviewers to Index

final_implicit_score_mat.set_index(final_implicit_score_mat.columns[0], inplace=True)
final_item_cf_score_mat.set_index(final_item_cf_score_mat.columns[0], inplace=True)
final_lightgcn_score_mat.set_index(final_lightgcn_score_mat.columns[0], inplace=True)
final_ncf_score_mat.set_index(final_ncf_score_mat.columns[0], inplace=True)
final_jobs_score_mat.set_index(final_jobs_score_mat.columns[0], inplace=True)
final_cb_score_mat.set_index(final_cb_score_mat.columns[0], inplace=True)
final_user_cf_score_mat.set_index(final_user_cf_score_mat.columns[0], inplace=True)


### Add back missing columns

missing_course_implicit = list(set(final_ncf_score_mat.columns) - set(final_implicit_score_mat.columns))
missing_course_jobs = list(set(final_ncf_score_mat.columns) - set(final_jobs_score_mat.columns))
missing_course_cb = list(set(final_ncf_score_mat.columns) - set(final_cb_score_mat.columns))
missing_course_user_cf = list(set(final_ncf_score_mat.columns) - set(final_user_cf_score_mat.columns))

for course in missing_course_implicit:
    final_implicit_score_mat[course] = 0

for course in missing_course_jobs:
    final_jobs_score_mat[course] = 0

for course in missing_course_cb:
    final_cb_score_mat[course] = 0

for course in missing_course_user_cf:
    final_user_cf_score_mat[course] = 0

### Weights to try

weights = {
    "implicit": [1/7, 1/12, 0.282],
    "ncf": [1/7, 1/12, 0.087],
    "lightgcn": [1/7, 1/12, 0.158],
    "jobs": [1/7, 1/2, 0.057],
    "item_cf": [1/7, 1/12, 0.091],
    "user_cf": [1/7, 1/12, 0.123],
    "cb": [1/7, 1/12, 0.203]
}

### Evaluate the function

def evaluate_recommendations(recommendations, actual):
    """
    Evaluate recommendations using precision and recall for named reviewers.

    :param recommendations: Pandas Series where index is reviewer names and values are lists of recommended courses.
    :param actual: Pandas Series where index is reviewer names and values are lists of actual courses.
    :returns: Tuple containing average precision and recall for all reviewers.
    """
    precision_list = []
    recall_list = []

    # Iterate over each reviewer's recommendations and actual courses
    for reviewer, recs in recommendations.items():
        acts = actual.get(reviewer, [])
        true_positives = set(recs) & set(acts)

        precision = len(true_positives) / len(recs) if recs else 0
        recall = len(true_positives) / len(acts) if acts else 0

        precision_list.append(precision)
        recall_list.append(recall)

    average_precision = sum(precision_list) / len(precision_list) if precision_list else 0
    average_recall = sum(recall_list) / len(recall_list) if recall_list else 0

    f1 = 2 * (average_precision * average_recall) / (average_precision + average_recall) if average_precision + average_recall else 0

    return average_precision, average_recall, f1

### Evaluate the model

final = pd.DataFrame(index=final_implicit_score_mat.index, columns=final_implicit_score_mat.columns)
courses_taken = train.groupby('Reviewer')['Course Name'].agg(list)
user_courses_map = courses_taken.to_dict()
best_weight_index = 0
best_f1 = 0
best_precision = 0
best_recall = 0

actual_courses = eval.groupby('Reviewer')['Course Name'].agg(list)

for i in range(3):
  for reviewer in final_implicit_score_mat.index:
    for course in final_implicit_score_mat.columns:
      final.at[reviewer, course] = list(weights.values())[0][i] * final_implicit_score_mat.at[reviewer, course]+ list(weights.values())[1][i] * final_ncf_score_mat.at[reviewer, course]+ list(weights.values())[2][i] * final_lightgcn_score_mat.at[reviewer, course] + list(weights.values())[3][i] * final_jobs_score_mat.at[reviewer, course] + list(weights.values())[4][i] * final_item_cf_score_mat.at[reviewer, course] + list(weights.values())[5][i] * final_user_cf_score_mat.at[reviewer, course] + list(weights.values())[6][i] * final_cb_score_mat.at[reviewer, course]

  for reviewer, courses in user_courses_map.items():
    if reviewer in final.index:
        for course in courses:
            if course in final.columns:
                final.at[reviewer, course] = 0
  final = final.astype(float)
  top_n_recs =  final.apply(lambda x: x.nlargest(10).index.tolist(), axis=1)
  precision, recall, f1 = evaluate_recommendations(top_n_recs, actual_courses)
  if f1 > best_f1:
    best_f1 = f1
    best_precision = precision
    best_recall = recall
    best_weight_index = i

best_weights = {key: values[best_weight_index] for key, values in weights.items()}
print("Best Weights (Evaluation): ", best_weights)
print("Best F1 Score (Evaluation): ", best_f1)
print("Best Precision (Evaluation): ", best_precision)
print("Best Recall (Evaluation): ", best_recall)

final

### Evaluation on Test Set

final_weights = {
    "implicit": list(weights.values())[0][best_weight_index],
    "ncf": list(weights.values())[1][best_weight_index],
    "lightgcn": list(weights.values())[2][best_weight_index],
    "jobs": list(weights.values())[3][best_weight_index],
    "item_cf": list(weights.values())[4][best_weight_index],
    "user_cf": list(weights.values())[5][best_weight_index],
    "cb": list(weights.values())[6][best_weight_index]
}

final = pd.DataFrame(index=final_implicit_score_mat.index, columns=final_implicit_score_mat.columns)
test_weight_index = 0
test_f1 = 0
test_precision = 0
test_recall = 0
top_n_rec = False
test_courses = test.groupby('Reviewer')['Course Name'].agg(list)

for reviewer in final_implicit_score_mat.index:
  for course in final_implicit_score_mat.columns:
    final.at[reviewer, course] = list(final_weights.values())[0] * final_implicit_score_mat.at[reviewer, course]+ list(final_weights.values())[1] * final_ncf_score_mat.at[reviewer, course]+ list(final_weights.values())[2] * final_lightgcn_score_mat.at[reviewer, course] + list(final_weights.values())[3] * final_jobs_score_mat.at[reviewer, course] + list(final_weights.values())[4] * final_item_cf_score_mat.at[reviewer, course] + list(final_weights.values())[5] * final_user_cf_score_mat.at[reviewer, course] + list(final_weights.values())[6] * final_cb_score_mat.at[reviewer, course]

for reviewer, courses in user_courses_map.items():
  if reviewer in final.index:
      for course in courses:
          if course in final.columns:
              final.at[reviewer, course] = 0
final = final.astype(float)
top_n_recs =  final.apply(lambda x: x.nlargest(10).index.tolist(), axis=1)
precision, recall, f1 = evaluate_recommendations(top_n_recs, test_courses)
test_f1 = f1
test_precision = precision
test_recall = recall
test_weight_index = i

test_weights = {key: values for key, values in final_weights.items()}
print("Best Weights (Test): ", test_weights)
print("Best F1 Score (Test): ", test_f1)
print("Best Precision (Test): ", test_precision)
print("Best Recall (Test): ", test_recall)

### Sample Recommendations

sample_user = "kumar k"

sample_user_recs = top_n_recs[sample_user]
print(f"Course taken by {sample_user}:\n{user_courses_map[sample_user]}")
print(f"Top 10 Recommendations for {sample_user}:")

for recc in  sample_user_recs[:10]:
    print(recc)
